restructure app-component; add authed component to handle signed in users
