+++ src/app/authorized/authorized.component.html
@@ -1,0 +1,44 @@
+<mat-toolbar class="toolbar" color="primary">
+
+  <img src="../assets/wood.png" id="log-logo">
+  <span class="text-logo">log</span>
+  
+  <span class="toolbar-spacer"></span>
+
+  <div class="log-selector-div" *ngIf="!showDashboard">
+    <mat-form-field style="text-align: right;" appearance="none">
+      <!-- <mat-label>{{ selectedLog }}</mat-label> -->
+      <mat-select class="log-select" [placeholder]="selectedLog">
+       <mat-option (click)="toDashboard()">
+         All Logs
+       </mat-option>
+       <mat-option (click)="openDialog()">
+         New Log
+       </mat-option>
+       <hr>
+       <label id="your-logs">Your Logs</label>
+
+       <mat-option *ngFor="let log of logs | async" [value]="log.id" (click)="onLogSelectionChange(log)"
+       [selected]="selectedLog == log.id">
+          {{ log.id }}
+       </mat-option>
+     </mat-select>
+   </mat-form-field>
+  </div>
+
+  <ngx-auth-firebaseui-avatar (onSignOut)="userSignedOut()"></ngx-auth-firebaseui-avatar>
+  
+</mat-toolbar> 
+
+<div *ngIf="showDashboard; else main">
+    <div id="initial-log-select">
+        <app-dashboard (hasSelected)="userSelected($event)"></app-dashboard>
+    </div>
+</div>
+
+<ng-template #main>
+    <div class="main-div">
+        <h2>{{ selectedLog }}</h2>
+        <app-log-table></app-log-table>
+    </div>
+</ng-template>
