--- src/app/sidebar/sidebar.component.css
@@ -1,55 +1,0 @@
-.sidenav-container {
-
-  height: 100%;
-}
-
-.sidenav {
-  padding-top: 60px;
-  width: 250px;
-  position: fixed;
-}
-
-.sidenav .mat-toolbar {
-  background: inherit;
-}
-
-/* .mat-toolbar.mat-primary {
-  position: sticky;
-  top: 0;
-  z-index: 10;
-} */
-
-
-.text-logo {
-  font-size:xx-large;
-  font-family:monospace;
-}
-
-#welcome-banner {
-  font-family: monospace;
-}
-
-.log-table-div {
-  padding: 50px;
-  padding-top: 90px;
-}
-
-.toolbar-spacer {
-  flex: 1 1 auto;
-}
-
-#login-register {
-  align-items: center;
-  width: min-content;
-}
-
-#your-logs {
-  font-size: larger;
-}
-
-#initial-log-select {
-  padding-top: 25px;
-  padding-left: 15%;
-  padding-right: 15%;
-}
-
