--- src/app/sidebar/sidebar.component.html
@@ -1,61 +1,0 @@
-<mat-sidenav-container class="sidenav-container">
-
-  <mat-sidenav #drawer class="sidenav"
-      [attr.role]="(isHandset$ | async) ? 'dialog' : 'navigation'"
-      [mode]="(isHandset$ | async) ? 'over' : 'side'"
-      [opened]="(isHandset$ | async) === false">
-
-      <mat-list *ngIf="authService.auth.user | async as user">
-        <mat-list-item id="welcome-banner">Hello, {{ user.displayName }}</mat-list-item>
-        <mat-divider></mat-divider>
-        
-        <mat-list-item><span id="your-logs">Logs</span>
-          <span class="toolbar-spacer"></span>
-          <button mat-icon-button color="primary" aria-label="" (click)="openDialog()">
-            <i class="material-icons">add_circle_outline</i>
-          </button>
-        </mat-list-item>
-        
-        <mat-list-item style="padding-left: 35%;" *ngIf="showSpinner"><mat-spinner [diameter]="30"></mat-spinner> </mat-list-item>
-                
-        <mat-accordion *ngIf="!showSpinner">
-          <mat-expansion-panel (open)="panelOpenState = true"
-                               (closed)="panelOpenState = false">
-            <mat-expansion-panel-header>
-              <mat-panel-title>
-                {{ selectedLog }}
-              </mat-panel-title>
-            </mat-expansion-panel-header>
-
-            <mat-divider></mat-divider>
-
-            <mat-selection-list #logSelector [multiple]="false">
-              <mat-list-option 
-                *ngFor="let log of logs | async" [value]="log.id"
-                (click)="onLogSelectionChange(log)"
-                [selected]="selectedLog == log.id">
-                {{ log.id }}
-              </mat-list-option>
-            </mat-selection-list>
-            <!-- <div *ngFor="let item of log.attributes | keyvalue">
-              <small>{{item.key}}: {{item.value}}</small>
-            </div> -->
-          </mat-expansion-panel>
-        </mat-accordion>
-
-        <mat-divider></mat-divider>
-    
-      </mat-list>
-
-  </mat-sidenav>
-
-  <mat-sidenav-content class="sidenav-content">
-
-    <!-- Main content area -->
-    <div class="log-table-div">
-      <app-log-table></app-log-table>
-    </div>
-
-  </mat-sidenav-content>
-</mat-sidenav-container>
-
